Add basic way to play from level
We're about halfway there with this. Three main issues. First, at the moment it requires that you've played from menu once before in order to set up the players. Second, it still relies on manually setting script execution order. Third, the object has to be manually enabled and disabled, instead of just detecting where we came from.

But for now I think it's good enough.